update requirements, bicep, 6 Speech samples
